create skeleton for information section
Here I finally realized the intention was to add my own content and
images to the webpage as my own choice as opposed to replicating the
nonsensical placeholders provided in the example.  As a result I
created a skeleton for the structure of the content I will add, to be
modified when I decide the focus of my project's content.